// src/app/business-logic/model/tasks/paramsItem.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface ParamsItem {
    /**
     * Section that the parameter belongs to
     */
    section?: string;
    /**
     * Name of the parameter. The combination of section and name should be unique
     */
    name?: string;
    /**
     * Value of the parameter
     */
    value?: string;
    /**
     * Type of the parameter. Optional
     */
    type?: string;
    /**
     * The parameter description. Optional
     */
    description?: string;
}